requirements.txt
streamlit
numpy==1.24.3
Pillow==9.5.0
scikit-image==0.21.0
matplotlib==3.7.1
opencv
